@@ -7,7 +7,7 @@ import { routes } from './app/app.routes';
 import { AppComponent } from './app/app.component';
 import { environment } from './environments/environment';
 
-import Mapboxgl from 'mapbox-gl'; // or "const mapboxgl = require('mapbox-gl');"
+import  Mapboxgl  from 'mapbox-gl'; // or "const mapboxgl = require('mapbox-gl');"
 import { provideHttpClient, withFetch } from '@angular/common/http';
 
 Mapboxgl.accessToken = 'pk.eyJ1Ijoic2Vic3RpaWFuIiwiYSI6ImNsdTl5eGV0NzBlaGsycXBubDlrd2xoZjIifQ.3cUWTjnCSnAmVUW--A9NDA';
