@@ -1,8 +1,8 @@
-import { CapacitorConfig } from '@capacitor/cli';
+import type { CapacitorConfig } from '@capacitor/cli';
 
 const config: CapacitorConfig = {
   appId: 'io.ionic.starter',
-  appName: 'Zippy App',
+  appName: 'zippyApp',
   webDir: 'www',
   server: {
     androidScheme: 'https'
