@@ -1,14 +1,12 @@
 import { Component } from '@angular/core';
 import { IonApp, IonRouterOutlet } from '@ionic/angular/standalone';
 
-
 @Component({
   selector: 'app-root',
   templateUrl: 'app.component.html',
   standalone: true,
   imports: [IonApp, IonRouterOutlet],
 })
-export class AppComponent  {
+export class AppComponent {
   constructor() {}
-
 }
